requirements.txt
Flask==3.0.3
python-telegram-bot==20.7
Telethon==1.34.0
python-dotenv==1.0.1
cryptography==42.0.5
APScheduler==3.10.4
requests==2.32.3